Split PackageInfo and loadPackageInfoFromFile into separate files.
