+++ src/index.ts
@@ -3,7 +3,6 @@ import path from 'path'
 import prompts from 'prompts-ncu'
 import spawn from 'spawn-please'
 import { cliOptionsMap } from './cli-options'
-import { PackageInfo } from './lib/PackageInfo'
 import { cacheClear } from './lib/cache'
 import chalk, { chalkInit } from './lib/chalk'
 import doctor from './lib/doctor'
@@ -21,6 +20,7 @@ import packageManagers from './package-managers'
 import { Index } from './types/IndexType'
 import { Options } from './types/Options'
 import { PackageFile } from './types/PackageFile'
+import { PackageInfo } from './types/PackageInfo'
 import { RunOptions } from './types/RunOptions'
 import { VersionSpec } from './types/VersionSpec'
